fix: anx-paragraph bug with element type and anx-title inside
by default, the anx-paragraph has a p tag. When adding a title to the
paragraph, the title was rendered inside the p tag. In this case, the p
tag is automatically changed to a div by the browser. This throws an
error in nuxt. (Client side does not match server side rendered)